// src/config/Types/config.d.ts
export interface TopLevelConfig {
  MONGODB_SCHPAY_URL: string;
  CENTRAL_SCH_PORT: string;
  secret: string;
  loginPassword: string;
  phoneNumber: string;
  portalCard: string;
  role: string;
  poolSource: string;
  firstPasswordSet: string;
  fullName: string;
  salt: string;
  ENCRYPTION_KEY: string;
  expiresIn: string;
  TEMP_TOKEN_SECRET: string;
  realm: string;
  roleId: string;
}